allow used-defined filetype-specific snippets
For each folder in `g:snippets_dir`, load filetype-specific snippets.

This allows users to define filetype-specific snippets, ontop of the the
ones defined by default. For example, new snippets can be defined for
php by placing a folder named 'php.snippets' in the `snippets` folder
inside any of the folders specified in `g:snippets_dir`.

`g:snippets_dir` is interpreted as a comma-separated list of folders.